[level 0] Title: 구슬을 나누는 경우의 수, Time: 0.04 ms, Memory: 70.6 MB -BaekjoonHub

+++ 프로그래머스/0/120840. 구슬을 나누는 경우의 수/구슬을 나누는 경우의 수.java
@@ -1,11 +1,11 @@
 class Solution {
-    public int solution(int balls, int share) {
+    public long solution(int balls, int share) {
         long answer = 1;
         int cnt = 1;
         while(share>=cnt){
             answer *= balls--;
             answer /= cnt++;
         }
-        return (int)answer;
+        return answer;
     }
 }
